An implementation of the Prototype pattern as shown on Source Making at
https://sourcemaking.com/design_patterns/prototype.

This code example was modeled off of this one

https://sourcemaking.com/design_patterns/prototype/java/1

I was interested in whether the prototype pattern could help me with a current
design choice I'm faced with at work. In the short term I'm not entirely
sure it's what's needed. I've thought about it and I think there is a better
way. We'll see on Monday.

However, I like the Prototype pattern... and this implementation makes use
of a Factory Registry to snag an instance of a Prototype.

Implementing it has helped gain a deeper understanding.